refator: use react navigation for reverse image screens
