+++ src/app/Scenes/ReverseImage/types.ts
@@ -1,0 +1,7 @@
+// tslint:disable-next-line:interface-over-type-literal
+export type ReverseImageNavigationStack = {
+  Camera: undefined
+  MultipleResults: {
+    artworkIDs: string[]
+  }
+}
